module_3/lessons/lesson_29_sql_basics/nyc-taxi-pipeline/requirements.txt
# Core
duckdb>=0.10.0
psycopg2-binary>=2.9.9
neo4j>=5.18.0
pandas>=2.2.0
python-dotenv>=1.0.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Dashboard
dash>=2.17.0
dash-bootstrap-components>=1.6.0
plotly>=5.22.0
requests>=2.31.0
